Soften the requirement to use Blobs in the body
A Str is really convenient and for someone just playing around with a basic
P6SGI application and for demonstration purposes, Str is really much nicer. So,
let's allow them while recommend against their use and providing requirements
and suggestions for how the app server is supposed to deal with them.