Heavily revising 0.3.Draft
After taking a hiatus from working on this for a bit, I've decided that defining
"canonical form" and "standard variations" is confusing and dumb. It would be
nice to have an ultimate form that was all that was required as far as
middleware and servers were concerned, but I think there's just too much value
in the simple forms. Therefore, I'm making all "standard forms" required and
making the "canonical form" just the "best, most flexible form".

Next, I have written some (untested) middleware to provide a way for any
middleware to support this best form without having to worry about the others.
Sadly, Rakudo does not implement roles and traits yet in a way that would make
this pretty and extensible, but we'll make do with this to start. I put some of
my other prettier and extensible attempts in comments for the moment.