Change the body message to allow Cool too
So, just allowing Str or Blob to be emitted by the body is not very Perlish.
Allow any Cool and specify that such values must be stringified and encoded. We
leave it to the servers to decide if they want to stringify and encode Any,
which might be reasonable or might just be scary.