Adding a special Application Execution section
Following dicussion with LeoNerd and john napiorkowski and others, I want to
make it clear that servers should be implemented in such a way as to maximize
the concurrency between application and server. This means the server should
start the application as soon as possible and return responses from the
application to the origin as soon as possible.