Huge update to setup configuration/runtime routines
I was very dissatisfied with the previous change, so I completely redid it. The
last just was not well-thought through and had far reaching consequences,
especially making middleware nearly impossible to implement. This change
increases the complexity of implementing servers and middleware to some extent
(a reasonable extent, in my judgement), but does not greatly increase the
complexity of applications.

I hope to hold the specification as still as possible from here and focus on
implementing the reference implementation instead.